+++ NEWS.md
@@ -30,6 +30,9 @@ Add fnction `util_paralogistic_stats_tbl()` to create a summary table of the par
 10. Fix #477 - Add function `util_inverse_weibull_param_estimate()` to estimate the parameters of the Inverse Weibull distribution.
 Add function `util_inverse_weibull_aic()` to calculate the AIC for the Inverse Weibull distribution.
 Add function `util_inverse_weibull_stats_tbl()` to create a summary table of the Inverse Weibull distribution.
+11. Fix #476 - Add function `util_inverse_pareto_param_estimate()` to estimate the parameters of the Inverse Pareto distribution.
+Add function `util_inverse_pareto_aic()` to calculate the AIC for the Inverse Pareto distribution.
+Add Function `util_inverse_pareto_stats_tbl()` to create a summary table of the Inverse Pareto distribution.
 
 ## Minor Improvements and Fixes
 1. Fix #468 - Update `util_negative_binomial_param_estimate()` to add the use of